[Gardening] Extract basic source info structs from RawComment.h
