+++ include/swift/AST/Module.h
@@ -23,8 +23,8 @@
 #include "swift/AST/Identifier.h"
 #include "swift/AST/Import.h"
 #include "swift/AST/LookupKinds.h"
-#include "swift/AST/RawComment.h"
 #include "swift/AST/Type.h"
+#include "swift/Basic/BasicSourceInfo.h"
 #include "swift/Basic/Compiler.h"
 #include "swift/Basic/OptionSet.h"
 #include "swift/Basic/STLExtras.h"
