+++ lib/AST/Module.cpp
@@ -880,12 +880,12 @@ SourceFile::getBasicLocsForDecl(const Decl *D) const {
 
   for (const auto &SRC : D->getRawComment(/*SerializedOK*/false).Comments) {
     Result.DocRanges.push_back(std::make_pair(
-      LineColumn { SRC.StartLine, SRC.StartColumn },
+      SourcePosition { SRC.StartLine, SRC.StartColumn },
       SRC.Range.getByteLength())
     );
   }
 
-  auto setLineColumn = [&SM](LineColumn &Home, SourceLoc Loc) {
+  auto setLineColumn = [&SM](SourcePosition &Home, SourceLoc Loc) {
     if (Loc.isValid()) {
       std::tie(Home.Line, Home.Column) = SM.getPresumedLineAndColumnForLoc(Loc);
     }
