+++ lib/AST/RawComment.cpp
@@ -262,43 +262,3 @@ CharSourceRange RawComment::getCharSourceRange() {
                 static_cast<const char *>(Start.getOpaquePointerValue());
   return CharSourceRange(Start, Length);
 }
-
-BasicSourceFileInfo::BasicSourceFileInfo(const SourceFile *SF)
-    : SFAndIsFromSF(SF, true) {
-  FilePath = SF->getFilename();
-}
-
-bool BasicSourceFileInfo::isFromSourceFile() const {
-  return SFAndIsFromSF.getInt();
-}
-
-void BasicSourceFileInfo::populateWithSourceFileIfNeeded() {
-  const auto *SF = SFAndIsFromSF.getPointer();
-  if (!SF)
-    return;
-  SWIFT_DEFER {
-    SFAndIsFromSF.setPointer(nullptr);
-  };
-
-  SourceManager &SM = SF->getASTContext().SourceMgr;
-
-  if (FilePath.empty())
-    return;
-  auto stat = SM.getFileSystem()->status(FilePath);
-  if (!stat)
-    return;
-
-  LastModified = stat->getLastModificationTime();
-  FileSize = stat->getSize();
-
-  if (SF->hasInterfaceHash()) {
-    InterfaceHashIncludingTypeMembers = SF->getInterfaceHashIncludingTypeMembers();
-    InterfaceHashExcludingTypeMembers = SF->getInterfaceHash();
-  } else {
-    // FIXME: Parse the file with EnableInterfaceHash option.
-    InterfaceHashIncludingTypeMembers = Fingerprint::ZERO();
-    InterfaceHashExcludingTypeMembers = Fingerprint::ZERO();
-  }
-
-  return;
-}
