+++ lib/Basic/CMakeLists.txt
@@ -42,6 +42,7 @@ generate_revision_inc(llvm_revision_inc LLVM "${LLVM_MAIN_SRC_DIR}")
 generate_revision_inc(swift_revision_inc Swift "${SWIFT_SOURCE_DIR}")
 
 add_swift_host_library(swiftBasic STATIC
+  BasicSourceInfo.cpp
   Cache.cpp
   ClusteredBitVector.cpp
   DiverseStack.cpp
