+++ lib/Serialization/ModuleFile.h
@@ -20,10 +20,10 @@
 #include "swift/AST/LinkLibrary.h"
 #include "swift/AST/FileUnit.h"
 #include "swift/AST/Module.h"
-#include "swift/AST/RawComment.h"
 #include "swift/AST/SILLayout.h"
-#include "swift/Serialization/Validation.h"
+#include "swift/Basic/BasicSourceInfo.h"
 #include "swift/Basic/LLVM.h"
+#include "swift/Serialization/Validation.h"
 #include "clang/AST/Type.h"
 #include "llvm/ADT/ArrayRef.h"
 #include "llvm/ADT/DenseMap.h"
