+++ lib/SymbolGraphGen/Symbol.cpp
@@ -14,6 +14,7 @@
 #include "swift/AST/Comment.h"
 #include "swift/AST/Module.h"
 #include "swift/AST/ParameterList.h"
+#include "swift/AST/RawComment.h"
 #include "swift/AST/USRGeneration.h"
 #include "swift/Basic/SourceManager.h"
 #include "AvailabilityMixin.h"
